apphub-service/apphub-service-skyxplore/apphub-service-skyxplore-data/src/main/java/com/github/saphyra/apphub/service/skyxplore/data/save_game/dao/system_connection/SystemConnectionModelValidator.java
package com.github.saphyra.apphub.service.skyxplore.data.save_game.dao.system_connection;

import com.github.saphyra.apphub.api.skyxplore.model.game.SystemConnectionModel;
import com.github.saphyra.apphub.lib.exception.ExceptionFactory;
import com.github.saphyra.apphub.service.skyxplore.data.save_game.dao.GameItemValidator;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Component;

import static java.util.Objects.isNull;

@Component
@RequiredArgsConstructor
@Slf4j
public class SystemConnectionModelValidator {
    private final GameItemValidator gameItemValidator;

    public void validate(SystemConnectionModel model) {
        gameItemValidator.validate(model);

        if (isNull(model.getLine())) {
            throw ExceptionFactory.invalidParam("line", "must not be null");
        }
    }
}
